♻️ WV-25 refactor : 컴포넌트 컨테이너화
